[refactor] relocate Home page
